+++ common/src/main/java/mixin/events/BlockItemEventMixin.java
@@ -8,7 +8,9 @@
 package dev.pandasystems.pandalib.mixin.events;
 
 import com.llamalad7.mixinextras.sugar.Local;
-import dev.pandasystems.pandalib.event.serverevents.BlockEvents;
+import dev.pandasystems.pandalib.event.client.ClientBlockEvents;
+import dev.pandasystems.pandalib.event.server.ServerBlockEvents;
+import dev.pandasystems.pandalib.platform.GameDataKt;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.context.BlockPlaceContext;
@@ -23,14 +25,29 @@ public class BlockItemEventMixin {
 	@Inject(method = "place", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/BlockItem;placeBlock(Lnet/minecraft/world/item/context/BlockPlaceContext;Lnet/minecraft/world/level/block/state/BlockState;)Z"), cancellable = true)
 	public void beforeBlockPlace(BlockPlaceContext context, CallbackInfoReturnable<InteractionResult> cir, @Local(ordinal = 1) BlockPlaceContext newContext) {
 		var blockState = newContext.getLevel().getBlockState(newContext.getClickedPos());
-		var cancelled = !BlockEvents.getBlockPlacePreEvent().getInvoker().invoke(newContext.getLevel(), newContext.getClickedPos(), blockState, newContext.getPlayer());
-		if (cancelled) {
-			cir.setReturnValue(InteractionResult.FAIL);
+		if (!GameDataKt.game.isHost()) {
+			ClientBlockEvents.placeEvent().getInvoker()
+					.invoke(newContext.getLevel(), newContext.getClickedPos(), blockState, newContext.getPlayer());
+		}
+
+		if (GameDataKt.game.isHost()) {
+			var cancelled = !ServerBlockEvents.placePreEvent().getInvoker()
+					.invoke(newContext.getLevel(), newContext.getClickedPos(), blockState, newContext.getPlayer());
+			if (cancelled) {
+				cir.setReturnValue(InteractionResult.FAIL);
+			}
 		}
 	}
 
 	@Inject(method = "placeBlock", at = @At("TAIL"))
 	public void afterBlockPlace(BlockPlaceContext context, BlockState state, CallbackInfoReturnable<Boolean> cir) {
-		BlockEvents.getBlockPlacePostEvent().getInvoker().invoke(context.getLevel(), context.getClickedPos(), state, context.getPlayer());
+		if (!GameDataKt.game.isHost()) {
+			ClientBlockEvents.placeEvent().getInvoker()
+					.invoke(context.getLevel(), context.getClickedPos(), state, context.getPlayer());
+		}
+		if (GameDataKt.game.isHost()) {
+			ServerBlockEvents.placePostEvent().getInvoker()
+					.invoke(context.getLevel(), context.getClickedPos(), state, context.getPlayer());
+		}
 	}
 }