+++ common/src/main/java/mixin/events/ServerPlayerGameModeEventMixin.java
@@ -7,7 +7,7 @@
 
 package dev.pandasystems.pandalib.mixin.events;
 
-import dev.pandasystems.pandalib.event.serverevents.BlockEvents;
+import dev.pandasystems.pandalib.event.server.ServerBlockEvents;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
@@ -27,7 +27,7 @@ public class ServerPlayerGameModeEventMixin {
 
 	@Inject(method = "destroyBlock", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/block/Block;playerWillDestroy(Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/entity/player/Player;)Lnet/minecraft/world/level/block/state/BlockState;"), cancellable = true)
 	public void onBlockBreakEvent(BlockPos pos, CallbackInfoReturnable<Boolean> cir) {
-		var cancelled = !BlockEvents.getBlockBreakPreEvent().getInvoker().invoke(this.level, pos, this.level.getBlockState(pos), this.player);
+		var cancelled = !ServerBlockEvents.breakPreEvent().getInvoker().invoke(this.level, pos, this.level.getBlockState(pos), this.player);
 		if (cancelled) {
 			cir.setReturnValue(false);
 		}
@@ -36,6 +36,6 @@ public class ServerPlayerGameModeEventMixin {
 	@Inject(method = "destroyBlock", at = @At("RETURN"))
 	public void onBlockBreakEventPost(BlockPos pos, CallbackInfoReturnable<Boolean> cir) {
 		if (!cir.getReturnValue()) return;
-		BlockEvents.getBlockBreakPostEvent().getInvoker().invoke(this.level, pos, this.level.getBlockState(pos), this.player);
+		ServerBlockEvents.breakPostEvent().getInvoker().invoke(this.level, pos, this.level.getBlockState(pos), this.player);
 	}
 }