+++ common/src/main/kotlin/networking/payloads/config/CommonConfigPayload.kt
@@ -20,7 +20,7 @@ import java.util.*
 data class CommonConfigPayload(
 	val resourceLocation: ResourceLocation,
 	val optionMap: Map<String, String>, // option path -> option JSON value in string
-	val ownerUuid: Optional<UUID>
+	val playerId: Optional<UUID>
 ) : CustomPacketPayload {
 	override fun type(): CustomPacketPayload.Type<CommonConfigPayload> = TYPE
 
@@ -29,7 +29,7 @@ data class CommonConfigPayload(
 		val CODEC: StreamCodec<FriendlyByteBuf, CommonConfigPayload> = StreamCodec.composite(
 			ResourceLocation.STREAM_CODEC, CommonConfigPayload::resourceLocation,
 			ByteBufCodecs.map(::Object2ObjectOpenHashMap, ByteBufCodecs.STRING_UTF8, ByteBufCodecs.STRING_UTF8), CommonConfigPayload::optionMap,
-			ByteBufCodecs.optional(UUIDUtil.STREAM_CODEC), CommonConfigPayload::ownerUuid,
+			ByteBufCodecs.optional(UUIDUtil.STREAM_CODEC), CommonConfigPayload::playerId,
 			::CommonConfigPayload
 		)
 	}