+++ common/src/main/kotlin/utils/Event.kt
@@ -12,40 +12,48 @@ import java.lang.invoke.MethodHandles
 import java.lang.reflect.Method
 import java.lang.reflect.Proxy
 
+@JvmOverloads
 @Suppress("UNCHECKED_CAST")
-fun <T : Function<R>, R> event(classType: Class<T>, combiner: (List<R>) -> R): Event<T, R> {
+fun <T : Function<R>, R> event(
+	classType: Class<T>,
+	combiner: (List<R>) -> R,
+	shouldContinue: (R) -> Boolean = { true }
+): Event<T> {
 	fun <T, R> invokeMethod(listener: T, method: Method, args: Array<*>): R {
 		return MethodHandles.lookup().unreflect(method)
 			.bindTo(listener).invokeWithArguments(*args) as R
 	}
 
 	return Event { listeners ->
-		Proxy.newProxyInstance(ClassLoader::class.java.classLoader, arrayOf(classType), object : AbstractInvocationHandler() {
+		Proxy.newProxyInstance(classType.classLoader ?: Thread.currentThread().contextClassLoader, arrayOf(classType), object : AbstractInvocationHandler() {
 			override fun handleInvocation(proxy: Any, method: Method, args: Array<*>): Any? {
-				val returnList = mutableListOf<Any>()
-				listeners.forEach { listener ->
-					val returnValue: Any = invokeMethod(listener, method, args)
+				val returnList = mutableListOf<R>()
+				for (listener in listeners) {
+					val returnValue = invokeMethod<T, R>(listener, method, args)
 					returnList += returnValue
+					if (!shouldContinue(returnValue)) {
+						break
+					}
 				}
-				return combiner(returnList as List<R>)
+				return combiner(returnList)
 			}
 		}) as T
 	}
 }
 
-inline fun <reified T : Function<Unit>> event(): Event<T, Unit> = event { }
+inline fun <reified T : Function<R>, R> event(
+	noinline shouldContinue: (R) -> Boolean = { true },
+	noinline combiner: (List<R>) -> R
+): Event<T> = event(T::class.java, combiner, shouldContinue)
 
-inline fun <reified T : Function<R>, R> event(noinline combiner: (List<R>) -> R): Event<T, R> = event(T::class.java, combiner)
+inline fun <reified T : Function<Unit>> event(): Event<T> = event { }
 
-inline fun <reified T : Function<R>, R> eventLastResult(defaultValue: R): Event<T, R> = event { results ->
-	results.lastOrNull() ?: defaultValue
-}
+inline fun <reified T : Function<R>, R> eventLastResult(defaultValue: R): Event<T> = event { results -> results.lastOrNull() ?: defaultValue }
 
-inline fun <reified T : Function<Boolean>> cancelableEvent(): Event<T, Boolean> = event { results ->
-	results.all { it }
-}
+inline fun <reified T : Function<Boolean>> cancelableEvent(shouldCancelRemaining: Boolean = true): Event<T> =
+	event({ if (shouldCancelRemaining) it else true }, { results -> results.all { it } })
 
-class Event<T : Function<R>, R>(private val function: (MutableList<T>) -> T) {
+class Event<T : Function<*>>(private val function: (MutableList<T>) -> T) {
 	private var _invoker: T? = null
 	private val listeners = mutableListOf<Pair<Int, T>>()